# WIP, ENH: more efficient hmaps