--- a/darshan-util/pydarshan/darshan/tests/test_heatmap_handling.py
+++ b/darshan-util/pydarshan/darshan/tests/test_heatmap_handling.py
@@ -332,24 +332,19 @@ def test_get_aggregate_data(log_file, expected_agg_data, mod, ops):
             "sample-dxt-simple.darshan",
             1,
             ["read", "write"],
-            np.array([[4040, 0, 0, 0, 0, 0, 0, 0, 0,
-                       0, 0, 0, 0, 0 , 0 ,0]]).reshape(16, 1),
+            np.array([[4040]]),
         ),
         (
             "sample-dxt-simple.darshan",
             4,
             ["read", "write"],
-            np.vstack((
             np.array([[0, 0, 0, 4040]]),
-            np.zeros((15, 4)))),
         ),
         (
             "sample-dxt-simple.darshan",
             10,
             ["read", "write"],
-            np.vstack((
             np.array([[0, 0, 0, 0, 0, 0, 0, 0, 0, 4040]]),
-            np.zeros((15, 10)))),
         ),
         # `dxt.darshan` is complex enough to warrant changing the
         # selected operations
@@ -571,7 +566,7 @@ def test_get_heatmap_df(
         # check the data is conserved
         assert actual_hmap_data.values.sum() == 4040
         # make sure the output array is the correct shape
-        assert actual_hmap_data.shape == (16, xbins)
+        assert actual_hmap_data.shape == (1, xbins)
         # make sure the output data contains identical values
         assert_allclose(actual_hmap_data.values, expected_hmap_data)