* even if we add a switch to disable DXT
data analysis for Python summary reports (see: gh-729),
runtime `HEATMAP` data can still explode the
memory footprint in cases where there are many
IO-inactive ranks, because the current Python
code adds explicit all-zero rows for the full
set of ranks

* this branch removes the requirement to have
explicit all-zero rows added for IO-inactive
ranks, for both DXT and runtime `HEATMAP`,
 and the performance improvements in our
`asv` suite below are clear, both on memory
footprint and timing (larger log files can have
other bottlenecks as well, and there are other PRs open
related to some of these)

* the caveat is that, although the testsuite
does pass, the new plotting strategy will
require some more work, so this is just to get
things rolling/WIP with TODO list similar to:

- [ ] adjust the new scatterplot approach to use
      some kind of custom markers to fit the bin
      sizes, instead of fixed-size scatter markers (or pivot
      to another suitable plotting strategy with all-zero
      rows absent)
- [ ] check the plotting/testing-related changes carefully; most
      likely comparing the HTML heatmap results for a large number
      of logs once things are more refined
- [ ] restore the colorbar label to the heatmap plots--apparently
      the presence of this label is not tested for in the suite

* adjust `test_get_heatmap_df` for more efficient single-rank
heatmap data structures (no more explicit all-zero ranks!)

* adjust `test_set_y_axis_ticks_and_labels` for new heatmap
plotting approach

* benchmarking with `asv continuous -e -b ".*heatmap.*" main treddy_heatmap_rank_efficiency`:

```
       before           after         ratio
     [3fb01f49]       [32397878]
     <main>           <treddy_heatmap_rank_efficiency>
-            134M             114M     0.85  dxt_heatmap.GetHeatMapDf.peakmem_get_heatmap_df(10000, 250, 0.01)
-            132M             111M     0.84  dxt_heatmap.GetHeatMapDf.peakmem_get_heatmap_df(10000, 250, 0.001)
-         800±2ms          576±2ms     0.72  dxt_heatmap.PlotDXTHeatMapSmall.time_plot_heatmap_builtin_logs('examples/example-logs/ior_hdf5_example.darshan', 1000)
-         799±5ms          574±2ms     0.72  dxt_heatmap.PlotDXTHeatMapSmall.time_plot_heatmap_builtin_logs('examples/example-logs/dxt.darshan', 1000)
-         780±4ms          549±1ms     0.70  dxt_heatmap.PlotDXTHeatMapSmall.time_plot_heatmap_builtin_logs('tests/input/sample-dxt-simple.darshan', 1000)
-         271±1ms        137±0.8ms     0.51  dxt_heatmap.PlotDXTHeatMapSmall.time_plot_heatmap_builtin_logs('examples/example-logs/dxt.darshan', 10)
-       353±0.6ms        178±0.4ms     0.51  dxt_heatmap.PlotDXTHeatMapSmall.time_plot_heatmap_builtin_logs('examples/example-logs/ior_hdf5_example.darshan', 100)
-       272±0.7ms        137±0.3ms     0.50  dxt_heatmap.PlotDXTHeatMapSmall.time_plot_heatmap_builtin_logs('examples/example-logs/ior_hdf5_example.darshan', 10)
-         354±2ms        178±0.4ms     0.50  dxt_heatmap.PlotDXTHeatMapSmall.time_plot_heatmap_builtin_logs('examples/example-logs/dxt.darshan', 100)
-         372±1ms        175±0.5ms     0.47  dxt_heatmap.PlotDXTHeatMapSmall.time_plot_heatmap_builtin_logs('tests/input/sample-dxt-simple.darshan', 100)
-       279±0.4ms        128±0.4ms     0.46  dxt_heatmap.PlotDXTHeatMapSmall.time_plot_heatmap_builtin_logs('tests/input/sample-dxt-simple.darshan', 10)

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.

```

To give reviewers an idea of the visual differences with the new plotting approach--two comparisons against the `main` branch are below. Maybe this is something I can come back to, or we could eventually find a student to drive forward a bit more. I think what we're seeing is pretty much entirely the result of using a fixed scatter marker "square" instead of filling areas based on widths of bins/ranks on the axes. That means that some data looks "too big" and other times "too small," sometimes simultaneously in both axis directions.

For `treddy_runtime_heatmap_inactive_ranks.darshan`, with this branch result followed by `main`:

![image](https://user-images.githubusercontent.com/7903078/188286718-29698075-1c1b-4860-9b41-a32517c4adeb.png)
![image](https://user-images.githubusercontent.com/7903078/188286724-244b59b2-b877-4ba6-ad63-a206f5f9503d.png)

For `snyder_acme.exe_id1253318_9-27-24239-1515303144625770178_2.darshan`, with this branch result followed by `main`:

![image](https://user-images.githubusercontent.com/7903078/188286739-60120fdc-7e2c-4922-8c61-b9e2f6b390e9.png)
![image](https://user-images.githubusercontent.com/7903078/188286754-5891f136-d2f8-45e1-9391-a00826a4f2f9.png)