@tylerjereddy 2022-09-12
The diff below the fold produces heatmaps that are perhaps a bit closer, but I need to dig a bit deeper on differences in more complex cases.

 <details>

```diff
--- a/darshan-util/pydarshan/darshan/experimental/plots/plot_dxt_heatmap.py
+++ b/darshan-util/pydarshan/darshan/experimental/plots/plot_dxt_heatmap.py
@@ -23,6 +23,7 @@ npt = MagicMock()
 
 import pandas as pd
 import seaborn as sns
+import matplotlib
 import matplotlib.pyplot as plt
 from matplotlib.colors import LogNorm
 
@@ -383,8 +384,6 @@ def plot_heatmap(
         hmap_df = heatmap_handling.get_heatmap_df(agg_df=agg_df, xbins=xbins, nprocs=nprocs)
     elif mod == "HEATMAP":
         hmap_df = report.heatmaps[submodule].to_df(ops=ops)
-        # mirror the DXT approach to heatmaps by
-        # adding all-zero rows for inactive ranks
         xbins = hmap_df.shape[1]
 
     # build the joint plot with marginal histograms
@@ -401,9 +400,27 @@ def plot_heatmap(
 
     x, y = np.meshgrid(np.arange(xbins),
                        np.asarray(hmap_df.index))
+    print("x:", x)
+    print("y:", y)
+    print("hmap_df:", hmap_df)
     # x and y both have shape (active_ranks, xbins)
     # rather than (nprocs, xbins)
-    hmap = jgrid.ax_joint.scatter(x, y, c=hmap_df, cmap="YlOrRd", norm=LogNorm(), marker="s")
+    print(hmap_df.columns)
+    bin_1 = hmap_df.columns[0]
+    tmax, runtime = determine_hmap_runtime(report=report)
+    bin_width = (x[0][1] - x[0][0])
+    print("bin_width:", bin_width)
+    bin_height = (xbins / nprocs) * bin_width
+    custom_path = [[0, 0],
+                   [0, bin_height],
+                   [bin_width, bin_height],
+                   [bin_width, 0],
+                   [0,0]]
+    marker_width_pixels, marker_height_pixels = jgrid.ax_joint.transData.transform((bin_width, bin_height))
+    size = marker_width_pixels * marker_height_pixels * (12/16) * (12/16) / 2
+    print("width (pixels):", marker_width_pixels)
+    print("height (pixels):", marker_height_pixels)
+    hmap = jgrid.ax_joint.scatter(x, y, c=hmap_df, cmap="YlOrRd", norm=LogNorm(), marker=custom_path, s=size)
     jgrid.ax_joint.set_ylim(0, nprocs)
     jgrid.fig.colorbar(hmap, ax=jgrid.ax_joint, orientation='vertical')
 
@@ -445,9 +462,9 @@ def plot_heatmap(
         align="edge",
     )
 
-    tmax, runtime = determine_hmap_runtime(report=report)
     # scale the x-axis to span the calculated run time
     xbin_max = xbins * (runtime / tmax)
+    print("xbin_max:", xbin_max)
     jgrid.ax_joint.set_xlim(0.0, xbin_max)
     # set the x and y tick locations and labels using the runtime
     set_x_axis_ticks_and_labels(jointgrid=jgrid, tmax=runtime, bin_max=xbin_max, n_xlabels=4)

```

</details>

![image](https://user-images.githubusercontent.com/7903078/189734018-8befa55b-d23f-47d7-8906-9e4178776339.png)

![image](https://user-images.githubusercontent.com/7903078/189734212-a78ae6d2-2acc-4980-8959-9f2e46358ea2.png)