@tylerjereddy 2022-09-12
If we gradually increase the DPI and/or figure dimensions of the `snyder_acme..` heatmap on `main` to check on the "ground truth" of where data is (taking screencaps of the higher res results, not actually pasting at that resolution), we do start to see more data showing up at higher DPI/larger size, so we should exercise some caution in what we target as "truth."

700 DPI:
![image](https://user-images.githubusercontent.com/7903078/189760955-fbe71c66-94ea-4854-923c-e1a877d5a592.png)

2000 DPI:
![image](https://user-images.githubusercontent.com/7903078/189761695-d66e1bee-46e8-4c3c-86d4-8dbd086941a5.png)

300 DPI, but expanding physical dimensions of the figure to (15, 30) inches:
![image](https://user-images.githubusercontent.com/7903078/189765134-2997e039-fe07-4141-b0a6-1d05dbdf866f.png)